+++ src/cli/runCli.js
@@ -1,8 +1,0 @@
-export async function runCli() {
-  throw new Error('src/cli/runCli is deprecated. Use bin/jwebgen.js as the runtime entrypoint.');
-}
-
-export function isHandledError(error) {
-  return !!error?.jwebgenHandled;
-}
-